add: show subject category, school_year, requisite
