+++ app/_types/syllabus.ts
@@ -4,3 +4,4 @@ import z from "zod"
 export type Syllabus = z.infer<typeof Syllabus>
 export type Subject = Syllabus["subjects"][number]
 export type SubjectCategory = Subject["category"]
+export type SubjectRequisite = Subject["requisite"]
